Improve methods of student service, add exception handling

+++ src/main/java/com/universityspa/service/people/StudentServiceImpl.java
@@ -2,13 +2,17 @@ package com.universityspa.service.people;
 
 import com.universityspa.dto.StudentDto;
 import com.universityspa.entity.Student;
+import com.universityspa.exception.NotFoundException;
 import com.universityspa.repository.StudentRepository;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.CrossOrigin;
+
+import java.util.List;
+import java.util.stream.Collectors;
 
 /**
  * Service class for Student.
@@ -21,7 +25,6 @@ import org.springframework.web.bind.annotation.CrossOrigin;
  */
 
 @Service
-@CrossOrigin(origins = "http://localhost:8081")
 public class StudentServiceImpl implements StudentService {
 
     @Autowired
@@ -33,66 +36,97 @@ public class StudentServiceImpl implements StudentService {
     /**
      * Method creates new student
      * It takes DTO, converts it to entity
-     * and returns entity
+     * save entity to the database, converts entity to dto
+     * and return dto-object
      *
      * @param studentDto
-     * @return savedStudent
+     * @return createdStudent
      */
     @Override
-    public Student addStudent(StudentDto studentDto) {
+    public StudentDto addStudent(StudentDto studentDto) {
         Student studentToCreate = convertToEntity(studentDto);
-
         Student savedStudent = studentRepository.saveAndFlush(studentToCreate);
-        return savedStudent;
+        StudentDto createdStudent = convertToDto(savedStudent);
+        return createdStudent;
     }
 
     /**
      * Method delete a student
      *
      * @param id of the student
+     * @throws NotFoundException if student doesn't exist
      */
     @Override
-    public void deleteStudent(Long id) {
+    public void deleteStudent(Long id) throws NotFoundException {
         Student deleteStudent = studentRepository.getOne(id);
-        studentRepository.delete(deleteStudent);
+        if (deleteStudent != null) {
+            studentRepository.delete(deleteStudent);
+        } else {
+            throw new NotFoundException("Unable to delete, student with such id doesn't exist");
+        }
     }
 
     /**
      * Method edits information of the student
      * It takes DTO, converts it to entity
-     * and returns entity
+     * save entity to the database, converts entity to dto
+     * and return dto-object
      *
+     * @param id of the student
      * @param studentDto
-     * @return savedStudent
+     * @return editedStudent
+     * @throws NotFoundException if student doesn't exist
      */
     @Override
-    public Student editStudent(StudentDto studentDto) {
-        Student studentToEdit = convertToEntity(studentDto);
+    public StudentDto editStudent(Long id, StudentDto studentDto) throws NotFoundException {
+        Student studentToEdit = studentRepository.getOne(id);
+        if (studentToEdit != null) {
+            studentToEdit = convertToEntity(studentDto);
+            Student savedStudent = studentRepository.saveAndFlush(studentToEdit);
+            StudentDto editedStudent = convertToDto(savedStudent);
+            return editedStudent;
+        } else {
+            throw new NotFoundException("Unable to edit, student with such id doesn't exist");
+        }
 
-        Student savedStudent = studentRepository.saveAndFlush(studentToEdit);
-        return savedStudent;
     }
 
     /**
-     * Method receives all students
+     * Method receives all students with pagination
      *
-     * @return [Student]
+     * @param pageable
+     * @return studentDtoPage
      */
     @Override
-    public Page<Student> getAll(Pageable pageable) {
-        return studentRepository.findAll(pageable);
+    public Page<StudentDto> getAll(Pageable pageable) {
+        Page<Student> studentPage = studentRepository.findAll(pageable);
+        int totalElements = (int) studentPage.getTotalElements();
+        List<StudentDto> studentDtoList = studentPage
+                .getContent()
+                .stream()
+                .map(student -> convertToDto(student))
+                .collect(Collectors.toList());
+
+        Page<StudentDto> studentDtoPage = new PageImpl<>(studentDtoList, pageable, totalElements);
+        return studentDtoPage;
     }
 
     /**
      * Method receives a student by id
      *
-     * @param id of the Student
-     * @return student
+     * @param id of the student
+     * @return studentDto
+     * @throws NotFoundException if student doesn't exist
      */
     @Override
-    public Student getById(Long id) {
+    public StudentDto getById(Long id) throws NotFoundException {
         Student student = studentRepository.getOne(id);
-        return student;
+        if (student != null) {
+            StudentDto studentDto = convertToDto(student);
+            return studentDto;
+        } else {
+            throw new NotFoundException("Student not found");
+        }
     }
 
     /**
@@ -112,13 +146,20 @@ public class StudentServiceImpl implements StudentService {
         return student;
     }
 
-//    @Override
-//    public StudentDto covertToDto(Student student) {
-//        StudentDto studentDto = modelMapper.map(student, StudentDto.class);
-//        studentDto.setFirstName(student.getFirstName());
-//        studentDto.setLastName(student.getLastName());
-//        studentDto.setCourse(student.getCourse());
-//        studentDto.setGroupId(student.getGroupId());
-//        return studentDto;
-//    }
+    /**
+     * Method of converting entity into the DTO
+     * Uses ModelMapper library
+     *
+     * @param student
+     * @return studentDto
+     */
+    @Override
+    public StudentDto convertToDto(Student student) {
+        StudentDto studentDto = modelMapper.map(student, StudentDto.class);
+        studentDto.setFirstName(student.getFirstName());
+        studentDto.setLastName(student.getLastName());
+        studentDto.setCourse(student.getCourse());
+        studentDto.setGroupId(student.getGroupId());
+        return studentDto;
+    }
 }
